+++ src/types/context.go
@@ -1,6 +1,10 @@
 package types
 
-import "os"
+import (
+	"fmt"
+	"os"
+	"path"
+)
 
 // Context represents contextual information about what application/service the user is running
 type Context struct {
@@ -9,26 +13,27 @@ type Context struct {
 	HasServiceContext bool
 }
 
-// GetContext returns a Context for the current working direcotry
-func GetContext() (Context, error) {
-	currentDir, err := os.Getwd()
-	if err != nil {
-		return Context{}, err
+// GetAppContext returns an AppContext for the application found at the passed location
+// by walkig up the directory tree until it finds an application.yml
+func GetAppContext(location string) (AppContext, error) {
+	walkDir := path.Clean(location)
+	if !path.IsAbs(location) {
+		return AppContext{}, fmt.Errorf("Could not get app context: location must be absolute")
 	}
-	appContext, err := GetAppContext(currentDir)
-	if err != nil {
-		return Context{}, err
+	for {
+		if _, err := os.Stat(path.Join(walkDir, "application.yml")); err == nil {
+			appConfig, err := NewAppConfig(walkDir)
+			if err != nil {
+				return AppContext{}, err
+			}
+			return AppContext{
+				Location: walkDir,
+				Config:   appConfig,
+			}, nil
+		}
+		if walkDir == "/" || walkDir == "." {
+			return AppContext{}, fmt.Errorf("%s is not an exosphere application", location)
+		}
+		walkDir = path.Dir(walkDir)
 	}
-	if _, err = os.Stat("service.yml"); err != nil {
-		return Context{AppContext: appContext}, nil
-	}
-	serviceContext, err := GetServiceContext(appContext, currentDir)
-	if err != nil {
-		return Context{}, err
-	}
-	return Context{
-		AppContext:        appContext,
-		ServiceContext:    serviceContext,
-		HasServiceContext: true,
-	}, nil
 }
