+++ src/types/service_context.go
@@ -1,12 +1,5 @@
 package types
 
-import (
-	"io/ioutil"
-	"path"
-
-	yaml "gopkg.in/yaml.v2"
-)
-
 // ServiceContext represents the exosphere service the user is running
 type ServiceContext struct {
 	Dir         string
@@ -14,22 +7,4 @@ type ServiceContext struct {
 	Config      ServiceConfig
 	AppContext  AppContext
 	ServiceData *ServiceData
-}
-
-// GetServiceContext returns a ServiceContext for the service found at the passed location
-func GetServiceContext(appContext AppContext, location string) (ServiceContext, error) {
-	serviceConfig := ServiceConfig{}
-	yamlFile, err := ioutil.ReadFile(path.Join(location, "service.yml"))
-	if err != nil {
-		return ServiceContext{}, err
-	}
-	if err = yaml.Unmarshal(yamlFile, &serviceConfig); err != nil {
-		return ServiceContext{}, err
-	}
-	return ServiceContext{
-		Dir:        path.Base(location),
-		Location:   location,
-		Config:     serviceConfig,
-		AppContext: appContext,
-	}, nil
 }
